requirements.txt
Django == 1.11
pytz == 2019.1
setuptools == 28.8.0
Pillow==6.0.0
djangorestframework==3.8.0
django-bootstrap-form

